Initial Commit
Adding base code. As of now, the script will read from a text file a list of domains to check. It will determine if the SSL for each domain is valid, expired, or within the given renewal window. If the domain is needing renewed or expired and email is sent with the status (Expired, Renew), and domain. Need to add exception to the logic statement to send an 'all clear/nothing needs done' email if nothing is detected.

All email configuration settings are loaded from a settings.py file that is located in the same directory as the script.